worker/pauseable refactor

+++ src/worker.ts
@@ -3,19 +3,38 @@ import Acceptable from './acceptable';
 import net from 'net';
 import cp from 'child_process';
 
-export default class Worker implements Acceptable {
+export interface Worker extends Acceptable {
+    reload() : void;
+    running() : boolean;
+    shutdown() : Promise<void>;
+}
+
+export default class WorkerProcess implements Worker {
     private entryPoint : string;
     private args : string[];
-    private worker : cp.ChildProcess;
+    private state : {
+        mode : 'accepting',
+        worker : cp.ChildProcess
+    } | {
+        mode : 'buffering',
+        buffer : net.Socket[]
+    };
 
     constructor(entryPoint : string, args : string[]) {
         this.entryPoint = entryPoint;
         this.args = args;
-        this.worker = this.fork();
+        this.state = { mode : 'accepting', worker : this.fork() }
     }
 
     accept(s : net.Socket) {
-        this.worker.send('connection', s);
+        switch (this.state.mode) {
+            case 'accepting':
+                this.state.worker.send('connection', s);
+                break;
+            case 'buffering':
+                this.state.buffer.push(s);
+                break;
+        }
     }
 
     private fork() {
@@ -26,21 +45,47 @@ export default class Worker implements Acceptable {
         return worker;
     }
 
+    running() {
+        return this.state.mode === 'accepting';
+    }
+
     reload() {
-        let worker = this.worker;
-        this.worker = this.fork();
+        let worker = this.fork();
+        switch (this.state.mode) {
+            case 'accepting':
+                // Reload is "rolling", we don't need to wait for this to end.
+                void this.shutdown();
+                break;
+            case 'buffering':
+                for (let s of this.state.buffer) {
+                    worker.send('connection', s);
+                }
+                break;
+        }
+        this.state = { mode : 'accepting', worker };
+    }
+
+    async shutdown() {
+        if (this.state.mode === 'buffering') {
+            return;
+        }
+
+        let worker = this.state.worker;
+        this.state = { mode : 'buffering', buffer : [] };
 
         worker.removeAllListeners('exit');
         worker.send('shutdown');
         worker.disconnect();
 
-        // If it doesn't respond to shutdown in 2 seconds, forcibly kill.
         let timeout = setTimeout(() => {
             worker.kill();
         }, 2000);
 
-        worker.on('exit', () => {
-            clearTimeout(timeout);
+        return new Promise<void>((resolve, reject) => {
+            worker.on('exit', () => {
+                clearTimeout(timeout);
+                resolve();
+            });
         });
     }
 }
